+++ packages/cubejs-backend-native/js/index.ts
@@ -1,6 +1,7 @@
 import fs from 'fs';
 import path from 'path';
 import { Writable } from 'stream';
+import { FileContent } from '@cubejs-backend/shared';
 // import { getEnv } from '@cubejs-backend/shared';
 
 export interface BaseMeta {
@@ -275,3 +276,22 @@ export const pythonLoadConfig = async (context: string, options: { file: string 
 
     return config;
 }
+
+export const initJinjaEngine = (options: { debugInfo?: boolean }): void => {
+    const native = loadNative();
+    return native.initJinjaEngine(options);
+};
+export const loadTemplate = (templateName: string, templateContent: string): void => {
+  const native = loadNative();
+  return native.loadTemplate(templateName, templateContent);
+};
+
+export const clearTemplates = (): void => {
+  const native = loadNative();
+  return native.clearTemplates();
+};
+
+export const renderTemplate = (templateName: string, context: unknown): string => {
+  const native = loadNative();
+  return native.renderTemplate(templateName, context);
+};