+++ packages/cubejs-backend-native/src/python/template/mod.rs
@@ -1,0 +1,4 @@
+mod entry;
+mod mj_value;
+
+pub use entry::template_register_module;